+++ internal/client/dataset_service.go
@@ -4,6 +4,18 @@ type CreateDatasetRequest struct {
 	DoSnapshot bool `json:"doSnapshot"`
 }
 
+type CreateDatasetFromURLRequest struct {
+	URL                       string             `json:"url"`
+	DoSnapshot                *bool              `json:"doSnapshot,omitempty"`
+	PersistDataAfterIngestion *bool              `json:"persistDataAfterIngestion,omitempty"`
+	SampleSize                *DatasetSampleSize `json:"sampleSize,omitempty"`
+}
+
+type DatasetSampleSize struct {
+	Type  string `json:"type"`
+	Value int64  `json:"value"`
+}
+
 type CreateDatasetResponse struct {
 	ID string `json:"datasetId"`
 }
@@ -33,4 +45,39 @@ type Dataset struct {
 	IsDataEngineEligible                  bool    `json:"isDataEngineEligible"`
 	IsWranglingEligible                   bool    `json:"isWranglingEligible"`
 	IsPlaygroundEvaluationDatasetEligible bool    `json:"isPlaygroundEvaluationDatasetEligible"`
+	Uri                                   string  `json:"uri"`
+}
+
+type CreateDatastoreRequest struct {
+	Type          string          `json:"type"`
+	CanonicalName string          `json:"canonicalName"`
+	Params        DatastoreParams `json:"params"`
+}
+
+type DatastoreParams struct {
+	Fields      []map[string]string `json:"fields"`
+	DriverID    *string             `json:"driverId,omitempty"`
+	JDBCURL     *string             `json:"jdbcUrl,omitempty"`
+	ConnectorID *string             `json:"connectorId,omitempty"`
+}
+
+type Datastore struct {
+	ID              string              `json:"id"`
+	CanonicalName   string              `json:"canonicalName"`
+	Type            string              `json:"type"`
+	DriverClassType *string             `json:"driverClassType,omitempty"`
+	Fields          []map[string]string `json:"fields,omitempty"`
+}
+
+type UpdateDatastoreRequest struct {
+	CanonicalName string          `json:"canonicalName"`
+	Params        DatastoreParams `json:"params"`
+}
+
+type TestDatastoreConnectionRequest struct {
+	CredentialID string `json:"credentialId"`
+}
+
+type TestDatastoreConnectionResponse struct {
+	Message string `json:"message"`
 }